pyproject.toml
[project]
name = "zvonilabackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard,standart]>=0.121.2",
    "jwt>=1.4.0",
    "psycopg[binary,pool]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "alembic>=1.17.2",
]
